+++ test/frontend/__fixtures__/event_categories.ts
@@ -1,0 +1,8 @@
+import { EventCategory } from 'models/event';
+
+const categories: Array<EventCategory> = [
+  { id: 1, name: 'Workshop' },
+  { id: 2, name: 'Internship' }
+];
+
+export default categories;