+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog Fab-manager
 
 - Fix a bug: members can't change/cancel their reservations
-- Fix a bug: admin events view defaults to the list tab
+- Fix a bug: admin events view should default to the list tab
+- Fix a bug: event creation form should not allow setting multiple times the same price category
 
 ## v5.7.2 2023 February 24